refactor: change directory location of useWindows and WindowsContext
