+++ src/view/features/popup/hooks/useAddTabToNewGroup.ts
@@ -2,7 +2,7 @@ import { useCallback, useContext } from "react";
 
 import { addToNewGroup } from "../../../../repository/TabsRepository";
 import { getWindows } from "../../../../repository/WindowsRepository";
-import { WindowsContext } from "../contexts/Windows";
+import { WindowsContext } from "../../../contexts/Windows";
 
 export const useAddTabToNewGroup = (): ((tabId: number) => Promise<void>) => {
   const { setWindows } = useContext(WindowsContext);
