+++ src/view/features/popup/hooks/useCloseAllTabs.ts
@@ -3,7 +3,7 @@ import { useCallback, useContext } from "react";
 import { Tab } from "../../../../model/Tab";
 import { closeAllTabs } from "../../../../repository/TabsRepository";
 import { getWindows } from "../../../../repository/WindowsRepository";
-import { WindowsContext } from "../contexts/Windows";
+import { WindowsContext } from "../../../contexts/Windows";
 
 export const useCloseAllTabs = (): ((tabs: Tab[]) => Promise<void>) => {
   const { setWindows } = useContext(WindowsContext);
