+++ src/view/features/popup/hooks/useExpandTabGroup.ts
@@ -2,7 +2,7 @@ import { useCallback, useContext } from "react";
 
 import { expandTabGroup } from "../../../../repository/TabGroupRepository";
 import { getWindows } from "../../../../repository/WindowsRepository";
-import { WindowsContext } from "../contexts/Windows";
+import { WindowsContext } from "../../../contexts/Windows";
 
 export const useExpandTabGroup = (): ((groupId: number) => Promise<void>) => {
   const { setWindows } = useContext(WindowsContext);
