+++ src/view/features/popup/hooks/useMoveTabGroupToOtherWindow.ts
@@ -3,7 +3,7 @@ import { useCallback, useContext } from "react";
 import { WindowId } from "../../../../model/Window";
 import { moveTabGroupToOtherWindow } from "../../../../repository/TabGroupRepository";
 import { getWindows } from "../../../../repository/WindowsRepository";
-import { WindowsContext } from "../contexts/Windows";
+import { WindowsContext } from "../../../contexts/Windows";
 
 export const useMoveTabGroupToOtherWindow = (): ((
   groupId: number,
