+++ src/view/features/popup/hooks/useUnpinAllTabs.ts
@@ -3,7 +3,7 @@ import { useCallback, useContext } from "react";
 import { Tab } from "../../../../model/Tab";
 import { unpinAllTabs } from "../../../../repository/TabsRepository";
 import { getWindows } from "../../../../repository/WindowsRepository";
-import { WindowsContext } from "../contexts/Windows";
+import { WindowsContext } from "../../../contexts/Windows";
 
 export const useUnpinAllTabs = (): ((tabs: Tab[]) => Promise<void>) => {
   const { setWindows } = useContext(WindowsContext);
