+++ src/view/features/popup/pages/App.tsx
@@ -2,10 +2,10 @@ import CssBaseline from "@mui/material/CssBaseline";
 import { useEffect } from "react";
 
 import { ThemeContext } from "../../../contexts/Theme";
+import { WindowsContext } from "../../../contexts/Windows";
 import { useTheme } from "../../../hooks/useTheme";
+import { useWindows } from "../../../hooks/useWindows";
 import PopupThemeProvider from "../components/ThemeProvider";
-import { WindowsContext } from "../contexts/Windows";
-import { useWindows } from "../hooks/useWindows";
 
 import Home from "./Home";
